requirements.txt
streamlit==1.36.0
google-generativeai==0.5.0
PyPDF2==3.0.1
pandas==2.2.2